# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Сезоны — API дашборда",
    "description": "Аналитика продаж ресторана из iiko.\n\nКонтракты: бэкенд отдаёт факты, представление — зона фронтенда. docs/frontend-handoff.md",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/integrations/iiko": {
      "get": {
        "tags": [
          "Integrations"
        ],
        "summary": "Настройки подключения iiko",
        "operationId": "getIikoIntegration",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IikoSettingsPublic"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Integrations"
        ],
        "summary": "Сохранить подключение iiko",
        "operationId": "putIikoIntegration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIikoSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IikoSettingsPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/integrations/iiko/sync": {
      "post": {
        "tags": [
          "Integrations"
        ],
        "summary": "Загрузить продажи и склад из iiko",
        "operationId": "syncIikoIntegration",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "full",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Full"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IikoSyncStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/foodcost/snapshot": {
      "get": {
        "tags": [
          "Фудкост"
        ],
        "summary": "Снимок фудкоста (1 RTT)",
        "operationId": "getFoodcostSnapshot",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 2000
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 12,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Foodcost"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/foodcost/totals": {
      "get": {
        "tags": [
          "Фудкост"
        ],
        "summary": "Get Totals",
        "operationId": "getFoodcostTotals",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 2000
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 12,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostTotals"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/foodcost/units": {
      "get": {
        "tags": [
          "Фудкост"
        ],
        "summary": "Get Units",
        "operationId": "getFoodcostUnits",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 2000
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 12,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UnitCost"
                  },
                  "title": "Response Getfoodcostunits"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/foodcost/groups": {
      "get": {
        "tags": [
          "Фудкост"
        ],
        "summary": "Get Groups",
        "operationId": "getFoodcostGroups",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 2000
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 12,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupCost"
                  },
                  "title": "Response Getfoodcostgroups"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/foodcost/products": {
      "get": {
        "tags": [
          "Фудкост"
        ],
        "summary": "Get Products",
        "operationId": "getFoodcostProducts",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 2000
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 12,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductCost"
                  },
                  "title": "Response Getfoodcostproducts"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CostTotals": {
        "properties": {
          "revenue": {
            "type": "number",
            "title": "Revenue",
            "description": "Выручка разреза за period (строки paid > 0)"
          },
          "cost": {
            "type": "number",
            "title": "Cost",
            "description": "Себестоимость фудкост-строк (paid > 0 и cost > 0): строки без техкарт в фудкосте не участвуют"
          },
          "revenueWithCost": {
            "type": "number",
            "title": "Revenuewithcost",
            "description": "Выручка фудкост-строк — знаменатель фудкоста; revenueWithCost / revenue = покрытие техкартами"
          },
          "prevRevenue": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prevrevenue",
            "description": "Те же три числа за compare-период; null — сравнивать не с чем (в compare нет платных чеков)"
          },
          "prevCost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prevcost",
            "description": "Себестоимость за compare"
          },
          "prevRevenueWithCost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prevrevenuewithcost",
            "description": "Выручка фудкост-строк за compare"
          },
          "goal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal",
            "description": "Цель фудкоста, %; null — цели не внесены (до модуля targets)"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revenue",
          "cost",
          "revenueWithCost",
          "prevRevenue",
          "prevCost",
          "prevRevenueWithCost",
          "goal"
        ],
        "title": "CostTotals",
        "description": "Тоталы страницы: факты + цель."
      },
...
      "Foodcost": {
        "properties": {
          "period": {
            "$ref": "#/components/schemas/src__schemas__base__Period",
            "description": "Показываемый период: месяц последнего закрытого дня (или ?year/?month)"
          },
          "compare": {
            "$ref": "#/components/schemas/src__schemas__base__Period",
            "description": "Период сравнения: непосредственно предшествующий период той же формы"
          },
          "totals": {
            "$ref": "#/components/schemas/CostTotals",
            "description": "Тоталы страницы («чистый фудкост» в сырье); равны сумме units по построению"
          },
          "dirty": {
            "type": "null",
            "title": "Dirty",
            "description": "Фудкост с учётом потерь — null до фазы 2: без списаний число было бы занижено"
          },
          "units": {
            "items": {
              "$ref": "#/components/schemas/UnitCost"
            },
            "type": "array",
            "title": "Units",
            "description": "Всегда четыре элемента: k, b, w, o (нулевые включены)"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/GroupCost"
            },
            "type": "array",
            "title": "Groups",
            "description": "Группы с продажами в period, по убыванию выручки; группа без продаж в этом периоде не отдаётся"
          },
          "products": {
            "items": {
              "$ref": "#/components/schemas/ProductCost"
            },
            "type": "array",
            "title": "Products",
            "description": "ВСЕ позиции с техкартами за period (без топ-N и порога шума — рейтинг и фильтры строит фронт), по убыванию выручки. Кормит диаграмму «Выгодность позиций»"
          },
          "discounts": {
            "$ref": "#/components/schemas/Discounts"
          },
          "losses": {
            "$ref": "#/components/schemas/Losses"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "period",
          "compare",
          "totals",
          "units",
          "groups",
          "products",
          "discounts",
          "losses"
        ],
        "title": "Foodcost"
      },
...
      "GroupCost": {
        "properties": {
          "revenue": {
            "type": "number",
            "title": "Revenue",
            "description": "Выручка разреза за period (строки paid > 0)"
          },
          "cost": {
            "type": "number",
            "title": "Cost",
            "description": "Себестоимость фудкост-строк (paid > 0 и cost > 0): строки без техкарт в фудкосте не участвуют"
          },
          "revenueWithCost": {
            "type": "number",
            "title": "Revenuewithcost",
            "description": "Выручка фудкост-строк — знаменатель фудкоста; revenueWithCost / revenue = покрытие техкартами"
          },
          "prevRevenue": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prevrevenue",
            "description": "Те же три числа за compare-период; null — сравнивать не с чем (в compare нет платных чеков)"
          },
          "prevCost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prevcost",
            "description": "Себестоимость за compare"
          },
          "prevRevenueWithCost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prevrevenuewithcost",
            "description": "Выручка фудкост-строк за compare"
          },
          "unit": {
            "type": "string",
            "enum": [
              "k",
              "b",
              "w",
              "o"
            ],
            "title": "Unit",
            "description": "Юнит группы"
          },
          "group": {
            "type": "string",
            "title": "Group",
            "description": "Группа iiko — папка, в которой лежит блюдо (живой разрез, терминология — frontend-handoff §0)"
          },
          "goal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal",
            "description": "Цель % (наследуется от юнита в Целях); null — нет цели"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revenue",
          "cost",
          "revenueWithCost",
          "prevRevenue",
          "prevCost",
          "prevRevenueWithCost",
          "unit",
          "group"
        ],
        "title": "GroupCost"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IikoSettingsPublic": {
        "properties": {
          "restaurant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Restaurant Id"
          },
          "configured": {
            "type": "boolean",
            "title": "Configured"
          },
          "iiko_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iiko Url"
          },
          "iiko_login": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iiko Login"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "sync": {
            "$ref": "#/components/schemas/IikoSyncPublic"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "restaurant_id",
          "configured",
          "sync"
        ],
        "title": "IikoSettingsPublic"
      },
...
      "IikoSyncStartResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "running",
            "title": "Status",
            "default": "running"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "started_at"
        ],
        "title": "IikoSyncStartResponse"
      },
...
      "ProductCost": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id",
            "description": "UUID блюда в iiko — стабильная идентичность позиции (имя меняется, id — никогда; как productId на складе). null — историческая строка без dish_id (база до пересоздания)"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Актуальное имя: из ПОСЛЕДНЕЙ продажи периода — переименованное блюдо не двоится, отдаётся одной строкой под новым именем (идентичность по id)"
          },
          "unit": {
            "type": "string",
            "enum": [
              "k",
              "b",
              "w",
              "o"
            ],
            "title": "Unit",
            "description": "Юнит позиции; 'o' — вне подразделений (в фильтрах юнитов на фронте не участвует, в режиме «Всё» — да)"
          },
          "qty": {
            "type": "number",
            "title": "Qty",
            "description": "Порций за период; дробное у весовых"
          },
          "revenue": {
            "type": "number",
            "title": "Revenue",
            "description": "Выручка фудкост-строк позиции (= её revenueWithCost: множество строк одно)"
          },
          "listValue": {
            "type": "number",
            "title": "Listvalue",
            "description": "Прейскурантная выручка тех же строк (Σ цен меню, БЕЗ скидок). Цена меню порции = listValue / qty — её показывать «Ценой» в тултипе (решение 17.07); скидки позиции = listValue − revenue. fc считать от revenue (фактическая экономика), НЕ от listValue"
          },
          "cost": {
            "type": "number",
            "title": "Cost",
            "description": "Себестоимость фудкост-строк позиции"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "unit",
          "qty",
          "revenue",
          "listValue",
          "cost"
        ],
        "title": "ProductCost",
        "description": "Позиция диаграммы «Выгодность позиций по фудкосту».\n\nТолько фудкост-строки (правило №3: paid > 0 и cost > 0) — позиция\nбез техкарты в рейтинге выгодности лгала бы (fc 0%). Производные —\nзона фронтенда: цена порции = revenue / qty, себестоимость порции =\ncost / qty, fc = cost / revenue; топ-N и порог шума — представление."
      },
...
      "UnitCost": {
        "properties": {
          "revenue": {
            "type": "number",
            "title": "Revenue",
            "description": "Выручка разреза за period (строки paid > 0)"
          },
          "cost": {
            "type": "number",
            "title": "Cost",
            "description": "Себестоимость фудкост-строк (paid > 0 и cost > 0): строки без техкарт в фудкосте не участвуют"
          },
          "revenueWithCost": {
            "type": "number",
            "title": "Revenuewithcost",
            "description": "Выручка фудкост-строк — знаменатель фудкоста; revenueWithCost / revenue = покрытие техкартами"
          },
          "prevRevenue": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prevrevenue",
            "description": "Те же три числа за compare-период; null — сравнивать не с чем (в compare нет платных чеков)"
          },
          "prevCost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prevcost",
            "description": "Себестоимость за compare"
          },
          "prevRevenueWithCost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prevrevenuewithcost",
            "description": "Выручка фудкост-строк за compare"
          },
          "key": {
            "type": "string",
            "enum": [
              "k",
              "b",
              "w",
              "o"
            ],
            "title": "Key",
            "description": "Юнит: k кухня, b бар, w вино, o «вне подразделений» (группы iiko вне папок Кухня/Бар/Вино)"
          },
          "goal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal",
            "description": "Цель фудкоста юнита, %; null — цель не задана в Целях"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revenue",
          "cost",
          "revenueWithCost",
          "prevRevenue",
          "prevCost",
          "prevRevenueWithCost",
          "key"
        ],
        "title": "UnitCost"
      },
...
      "UpdateIikoSettingsRequest": {
        "properties": {
          "iiko_url": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Iiko Url"
          },
          "iiko_login": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Iiko Login"
          },
          "iiko_password": {
            "type": "string",
            "maxLength": 200,
            "title": "Iiko Password",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "iiko_url",
          "iiko_login"
        ],
        "title": "UpdateIikoSettingsRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}